+++ packages/flutter_tools/test/general.shard/resident_web_runner_cold_test.dart
@@ -27,6 +27,7 @@ import 'package:vm_service/vm_service.dart';
 import 'package:webkit_inspection_protocol/webkit_inspection_protocol.dart';
 
 import '../src/common.dart';
+import '../src/fakes.dart';
 import '../src/testbed.dart';
 
 void main() {
@@ -58,7 +59,7 @@ void main() {
           urlTunneller: null,
         ) as ResidentWebRunner;
       }, overrides: <Type, Generator>{
-        Pub: () => MockPub(),
+        Pub: () => FakePub(),
       }
     );
   });
@@ -201,5 +202,4 @@ class MockChromeConnection extends Mock implements ChromeConnection {}
 class MockChromeTab extends Mock implements ChromeTab {}
 class MockWipConnection extends Mock implements WipConnection {}
 class MockBuildSystem extends Mock implements BuildSystem {}
-class MockPub extends Mock implements Pub {}
 class MockChromiumLauncher extends Mock implements ChromiumLauncher {}
